Remove test_dataframe.py
