+++ python/cudf/cudf/tests/dataframe/indexing/test_setitem.py
@@ -8,6 +8,10 @@ import pandas as pd
 import pytest
 
 import cudf
+from cudf.core._compat import (
+    PANDAS_CURRENT_SUPPORTED_VERSION,
+    PANDAS_VERSION,
+)
 from cudf.testing import assert_eq
 from cudf.testing._utils import assert_exceptions_equal, expect_warning_if
 
@@ -339,4 +343,429 @@ def test_multicolumn_set_item():
     gdf = cudf.from_pandas(pdf)
     pdf["d"] = [1, 2, 3, 4, 5]
     gdf["d"] = [1, 2, 3, 4, 5]
+    assert_eq(pdf, gdf)
+
+
+def test_dataframe_basic():
+    rng = np.random.default_rng(seed=0)
+    df = cudf.DataFrame()
+
+    # Populate with cuda memory
+    df["keys"] = np.arange(10, dtype=np.float64)
+    np.testing.assert_equal(df["keys"].to_numpy(), np.arange(10))
+    assert len(df) == 10
+
+    # Populate with numpy array
+    rnd_vals = rng.random(10)
+    df["vals"] = rnd_vals
+    np.testing.assert_equal(df["vals"].to_numpy(), rnd_vals)
+    assert len(df) == 10
+    assert tuple(df.columns) == ("keys", "vals")
+
+    # Make another dataframe
+    df2 = cudf.DataFrame()
+    df2["keys"] = np.array([123], dtype=np.float64)
+    df2["vals"] = np.array([321], dtype=np.float64)
+
+    # Concat
+    df = cudf.concat([df, df2])
+    assert len(df) == 11
+
+    hkeys = np.asarray([*np.arange(10, dtype=np.float64).tolist(), 123])
+    hvals = np.asarray([*rnd_vals.tolist(), 321])
+
+    np.testing.assert_equal(df["keys"].to_numpy(), hkeys)
+    np.testing.assert_equal(df["vals"].to_numpy(), hvals)
+
+    # As matrix
+    mat = df.values_host
+
+    expect = np.vstack([hkeys, hvals]).T
+
+    np.testing.assert_equal(mat, expect)
+
+    # test dataframe with tuple name
+    df_tup = cudf.DataFrame()
+    data = np.arange(10)
+    df_tup[(1, "foobar")] = data
+    np.testing.assert_equal(data, df_tup[(1, "foobar")].to_numpy())
+
+    df = cudf.DataFrame(pd.DataFrame({"a": [1, 2, 3], "c": ["a", "b", "c"]}))
+    pdf = pd.DataFrame(pd.DataFrame({"a": [1, 2, 3], "c": ["a", "b", "c"]}))
+    assert_eq(df, pdf)
+
+    gdf = cudf.DataFrame({"id": [0, 1], "val": [None, None]})
+    gdf["val"] = gdf["val"].astype("int")
+
+    assert gdf["val"].isnull().all()
+
+
+def test_dataframe_column_add_drop_via_setitem():
+    df = cudf.DataFrame()
+    data = np.asarray(range(10))
+    df["a"] = data
+    df["b"] = data
+    assert tuple(df.columns) == ("a", "b")
+    del df["a"]
+    assert tuple(df.columns) == ("b",)
+    df["c"] = data
+    assert tuple(df.columns) == ("b", "c")
+    df["a"] = data
+    assert tuple(df.columns) == ("b", "c", "a")
+
+
+def test_dataframe_column_set_via_attr():
+    data_0 = np.asarray([0, 2, 4, 5])
+    data_1 = np.asarray([1, 4, 2, 3])
+    data_2 = np.asarray([2, 0, 3, 0])
+    df = cudf.DataFrame({"a": data_0, "b": data_1, "c": data_2})
+
+    for i in range(10):
+        df.c = df.a
+        assert assert_eq(df.c, df.a, check_names=False)
+        assert tuple(df.columns) == ("a", "b", "c")
+
+        df.c = df.b
+        assert assert_eq(df.c, df.b, check_names=False)
+        assert tuple(df.columns) == ("a", "b", "c")
+
+
+@pytest.mark.parametrize("nelem", [0, 10])
+def test_dataframe_astype(nelem):
+    df = cudf.DataFrame()
+    data = np.asarray(range(nelem), dtype=np.int32)
+    df["a"] = data
+    assert df["a"].dtype is np.dtype(np.int32)
+    df["b"] = df["a"].astype(np.float32)
+    assert df["b"].dtype is np.dtype(np.float32)
+    np.testing.assert_equal(df["a"].to_numpy(), df["b"].to_numpy())
+
+
+def test_dataframe_add_col_to_object_dataframe():
+    # Test for adding column to an empty object dataframe
+    cols = ["a", "b", "c"]
+    df = pd.DataFrame(columns=cols, dtype="str")
+
+    data = {k: ["a"] for k in cols}
+
+    gdf = cudf.DataFrame(data)
+    gdf = gdf[:0]
+
+    assert gdf.dtypes.equals(df.dtypes)
+    gdf["a"] = [1]
+    df["a"] = [10]
+    assert gdf.dtypes.equals(df.dtypes)
+    gdf["b"] = [1.0]
+    df["b"] = [10.0]
+    assert gdf.dtypes.equals(df.dtypes)
+
+
+def test_dataframe_append_empty():
+    pdf = pd.DataFrame(
+        {
+            "key": [1, 1, 1, 2, 2, 2, 3, 3, 3, 4, 4, 4],
+            "value": [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12],
+        }
+    )
+    gdf = cudf.DataFrame.from_pandas(pdf)
+
+    gdf["newcol"] = 100
+    pdf["newcol"] = 100
+
+    assert len(gdf["newcol"]) == len(pdf)
+    assert len(pdf["newcol"]) == len(pdf)
+    assert_eq(gdf, pdf)
+
+
+def test_dataframe_setitem_from_masked_object():
+    rng = np.random.default_rng(seed=0)
+    ary = rng.standard_normal(100)
+    mask = np.zeros(100, dtype=bool)
+    mask[:20] = True
+    rng.shuffle(mask)
+    ary[mask] = np.nan
+
+    test1_null = cudf.Series(ary, nan_as_null=True)
+    assert test1_null.null_count == 20
+    test1_nan = cudf.Series(ary, nan_as_null=False)
+    assert test1_nan.null_count == 0
+
+    test2_null = cudf.DataFrame.from_pandas(
+        pd.DataFrame({"a": ary}), nan_as_null=True
+    )
+    assert test2_null["a"].null_count == 20
+    test2_nan = cudf.DataFrame.from_pandas(
+        pd.DataFrame({"a": ary}), nan_as_null=False
+    )
+    assert test2_nan["a"].null_count == 0
+
+    gpu_ary = cp.asarray(ary)
+    test3_null = cudf.Series(gpu_ary, nan_as_null=True)
+    assert test3_null.null_count == 20
+    test3_nan = cudf.Series(gpu_ary, nan_as_null=False)
+    assert test3_nan.null_count == 0
+
+    test4 = cudf.DataFrame()
+    lst = [1, 2, None, 4, 5, 6, None, 8, 9]
+    test4["lst"] = lst
+    assert test4["lst"].null_count == 2
+
+
+def test_dataframe_append_to_empty():
+    pdf = pd.DataFrame()
+    pdf["a"] = []
+    pdf["a"] = pdf["a"].astype("str")
+    pdf["b"] = [1, 2, 3]
+
+    gdf = cudf.DataFrame()
+    gdf["a"] = []
+    gdf["b"] = [1, 2, 3]
+
+    assert_eq(gdf, pdf)
+
+
+def test_dataframe_setitem_index_len1():
+    gdf = cudf.DataFrame()
+    gdf["a"] = [1]
+    gdf["b"] = gdf.index._column
+
+    np.testing.assert_equal(gdf.b.to_numpy(), [0])
+
+
+def test_empty_dataframe_setitem_df():
+    gdf1 = cudf.DataFrame()
+    gdf2 = cudf.DataFrame({"a": [1, 2, 3, 4, 5]})
+    gdf1["a"] = gdf2["a"]
+    assert_eq(gdf1, gdf2)
+
+
+@pytest.mark.parametrize("nrows", [0, 3])
+def test_nonmatching_index_setitem(nrows):
+    rng = np.random.default_rng(seed=0)
+
+    gdf = cudf.DataFrame()
+    gdf["a"] = rng.integers(2147483647, size=nrows)
+    gdf["b"] = rng.integers(2147483647, size=nrows)
+    gdf = gdf.set_index("b")
+
+    test_values = rng.integers(2147483647, size=nrows)
+    gdf["c"] = test_values
+    assert len(test_values) == len(gdf["c"])
+    gdf_series = cudf.Series(test_values, index=gdf.index, name="c")
+    assert_eq(gdf["c"].to_pandas(), gdf_series.to_pandas())
+
+
+def test_dataframe_assignment():
+    pdf = pd.DataFrame()
+    for col in "abc":
+        pdf[col] = np.array([0, 1, 1, -2, 10])
+    gdf = cudf.DataFrame.from_pandas(pdf)
+    gdf[gdf < 0] = 999
+    pdf[pdf < 0] = 999
+    assert_eq(gdf, pdf)
+
+
+@pytest.mark.parametrize("a", [[], ["123"]])
+@pytest.mark.parametrize("b", ["123", ["123"]])
+@pytest.mark.parametrize(
+    "misc_data",
+    ["123", ["123"] * 20, 123, [1, 2, 0.8, 0.9] * 50, 0.9, 0.00001],
+)
+@pytest.mark.parametrize("non_list_data", [123, "abc", "zyx", "rapids", 0.8])
+def test_create_dataframe_cols_empty_data(a, b, misc_data, non_list_data):
+    expected = pd.DataFrame({"a": a})
+    actual = cudf.DataFrame.from_pandas(expected)
+    expected["b"] = b
+    actual["b"] = b
+    assert_eq(actual, expected)
+
+    expected = pd.DataFrame({"a": []})
+    actual = cudf.DataFrame.from_pandas(expected)
+    expected["b"] = misc_data
+    actual["b"] = misc_data
+    assert_eq(actual, expected)
+
+    expected = pd.DataFrame({"a": a})
+    actual = cudf.DataFrame.from_pandas(expected)
+    expected["b"] = non_list_data
+    actual["b"] = non_list_data
+    assert_eq(actual, expected)
+
+
+@pytest.mark.parametrize(
+    "list_input",
+    [
+        pytest.param([1, 2, 3, 4], id="smaller"),
+        pytest.param([1, 2, 3, 4, 5, 6], id="larger"),
+    ],
+)
+@pytest.mark.parametrize(
+    "key",
+    [
+        pytest.param("list_test", id="new_column"),
+        pytest.param("id", id="existing_column"),
+    ],
+)
+def test_setitem_diff_size_list(list_input, key):
+    gdf = cudf.DataFrame(range(5))
+    with pytest.raises(
+        ValueError, match="All columns must be of equal length"
+    ):
+        gdf[key] = list_input
+
+
+@pytest.mark.parametrize(
+    "data, index",
+    [
+        [[1, 2, 3, 4], None],
+        [[1, 2, 3, 4, 5, 6], None],
+        [[1, 2, 3], [4, 5, 6]],
+    ],
+)
+@pytest.mark.parametrize("klass", [pd.Series, cudf.Series])
+@pytest.mark.parametrize(
+    "key",
+    [
+        pytest.param("list_test", id="new_column"),
+        pytest.param("id", id="existing_column"),
+    ],
+)
+def test_setitem_diff_size_series(klass, data, index, key):
+    gdf = cudf.DataFrame(range(5))
+    pdf = gdf.to_pandas()
+
+    series_input = klass(data, index=index)
+    pandas_input = series_input
+    if isinstance(pandas_input, cudf.Series):
+        pandas_input = pandas_input.to_pandas()
+
+    expect = pdf
+    expect[key] = pandas_input
+
+    got = gdf
+    got[key] = series_input
+
+    # Pandas uses NaN and typecasts to float64 if there's missing values on
+    # alignment, so need to typecast to float64 for equality comparison
+    expect = expect.astype("float64")
+    got = got.astype("float64")
+
+    assert_eq(expect, got)
+
+
+def test_tupleize_cols_False_set():
+    pdf = pd.DataFrame()
+    gdf = cudf.DataFrame()
+    pdf[("a", "b")] = [1]
+    gdf[("a", "b")] = [1]
+    assert_eq(pdf, gdf)
+    assert_eq(pdf.columns, gdf.columns)
+
+
+@pytest.mark.parametrize(
+    "col_data",
+    [
+        range(5),
+        ["a", "b", "x", "y", "z"],
+        [1.0, 0.213, 0.34332],
+        ["a"],
+        [1],
+        [0.2323],
+        [],
+    ],
+)
+@pytest.mark.parametrize(
+    "assign_val",
+    [
+        1,
+        2,
+        np.array(2),
+        cp.array(2),
+        0.32324,
+        np.array(0.34248),
+        cp.array(0.34248),
+        "abc",
+        np.array("abc", dtype="object"),
+        np.array("abc", dtype="str"),
+        np.array("abc"),
+        None,
+    ],
+)
+@pytest.mark.skipif(
+    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
+    reason="Fails in older versions of pandas",
+)
+def test_dataframe_assign_scalar(request, col_data, assign_val):
+    request.applymarker(
+        pytest.mark.xfail(
+            condition=PANDAS_VERSION >= PANDAS_CURRENT_SUPPORTED_VERSION
+            and len(col_data) == 0,
+            reason="https://github.com/pandas-dev/pandas/issues/56679",
+        )
+    )
+    pdf = pd.DataFrame({"a": col_data})
+    gdf = cudf.DataFrame({"a": col_data})
+
+    pdf["b"] = (
+        cp.asnumpy(assign_val)
+        if isinstance(assign_val, cp.ndarray)
+        else assign_val
+    )
+    gdf["b"] = assign_val
+
+    assert_eq(pdf, gdf)
+
+
+@pytest.mark.parametrize(
+    "col_data",
+    [
+        1,
+        2,
+        np.array(2),
+        cp.array(2),
+        0.32324,
+        np.array(0.34248),
+        cp.array(0.34248),
+        "abc",
+        np.array("abc", dtype="object"),
+        np.array("abc", dtype="str"),
+        np.array("abc"),
+        None,
+    ],
+)
+@pytest.mark.parametrize(
+    "assign_val",
+    [
+        1,
+        2,
+        np.array(2),
+        cp.array(2),
+        0.32324,
+        np.array(0.34248),
+        cp.array(0.34248),
+        "abc",
+        np.array("abc", dtype="object"),
+        np.array("abc", dtype="str"),
+        np.array("abc"),
+        None,
+    ],
+)
+def test_dataframe_assign_scalar_with_scalar_cols(col_data, assign_val):
+    pdf = pd.DataFrame(
+        {
+            "a": cp.asnumpy(col_data)
+            if isinstance(col_data, cp.ndarray)
+            else col_data
+        },
+        index=["dummy_mandatory_index"],
+    )
+    gdf = cudf.DataFrame({"a": col_data}, index=["dummy_mandatory_index"])
+
+    pdf["b"] = (
+        cp.asnumpy(assign_val)
+        if isinstance(assign_val, cp.ndarray)
+        else assign_val
+    )
+    gdf["b"] = assign_val
+
     assert_eq(pdf, gdf)
